requirements.txt
asgiref==3.8.1
Django==4.1.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
pymongo==3.12.3
pytz==2024.1
sqlparse==0.2.4
